server/requirements.txt
flask
flask-sqlalchemy
flask-migrate
sqlalchemy-serializer
flask-restful
python-dotenv
flask-cors
faker
Werkzeug
importlib-metadata
importlib-resources
